% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.R
\name{get_slideshow}
\alias{get_slideshow}
\title{Get Slideshow Information}
\usage{
get_slideshow(slideshow_url = "", username = auth_username(NULL),
  password = auth_password(NULL), exclude_tags = 1, detailed = 1,
  get_transcript = 1)
}
\arguments{
\item{slideshow_url}{- URL of presentation}

\item{exclude_tags}{- Exclude tags from the detailed information. 1 to exclude.}

\item{detailed}{- Whether or not to include optional information.
1 to include, 0 (default) for basic information.}

\item{get_transcript}{- 0 (default) or 1 to include}
}
\examples{
slideshow <- get_slideshow(slideshow_url = "http://www.slideshare.net/EdithYeung/8-tips-for-scaling-mobile-users-in-china-by-edith-yeung/")

}

