% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_user_contacts}
\alias{get_user_contacts}
\title{Returns user contacts}
\usage{
get_user_contacts(username_for = "", username = auth_username(NULL),
  password = auth_password(NULL))

get_user_contacts(username_for = "", username = auth_username(NULL),
  password = auth_password(NULL))
}
\examples{
asdqwdq <- get_user_contacts(username_for = "rashmi")

asdqwdq <- get_user_groups(username_for = "rashmi")

}

