% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_user_tags}
\alias{get_user_tags}
\title{Returns user tags}
\usage{
get_user_tags(username = auth_username(NULL),
  password = auth_password(NULL))
}
\examples{
asdqwdq <- get_user_tags(username = "f789gh", password = "")

}

