% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{userAPIauth}
\alias{userAPIauth}
\title{key: 7CECRDW0HnGsKmAJ8sjQ
secret: Z8Pk4MJ4WPX1BefN6exBjkhs9vd74UfTnxDqz7LfQ}
\usage{
userAPIauth()
}
\description{
key: 7CECRDW0HnGsKmAJ8sjQ
secret: Z8Pk4MJ4WPX1BefN6exBjkhs9vd74UfTnxDqz7LfQ
}

