% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_api_key}
\alias{auth_api_key}
\alias{auth_shared_secret}
\alias{authentication_slideshare}
\title{SlideShare authentication}
\usage{
auth_api_key(x)

auth_shared_secret(x)

authentication_slideshare()
}
\examples{
options(API_key = "nRWFyyZI", Shared_Secret = "dmvSBNkR")

}

