% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_user_contacts}
\alias{get_user_contacts}
\title{Returns user contacts}
\usage{
get_user_contacts(username_for = "", limit = NULL, offset = NULL)
}
\examples{
asdqwdq <- get_user_contacts(username_for = "rashmi")

}

